Amended schema to include uniqueness, a category field, and sample items generated inside server.js.
