+++ client/Views/postList.html
@@ -1,0 +1,9 @@
+<h1>POST PAGE</h1>
+<div ng-repeat='item in tests track by $index'>
+  <div><h2>POST</h2></div>
+  <div>{{ item.username }}</div>
+  <div>{{ item.title }}</div>
+  <div>{{ item.message }}</div>
+  <div>{{ item.category }}</div>
+  <div>{{ item.votes }}</div>
+</div>
