+++ client/index.html
@@ -10,8 +10,9 @@
     <div ui-view>
     </div>
     <script src="https://ajax.googleapis.com/ajax/libs/angularjs/1.4.5/angular.min.js"></script>
-    <script src="node_modules/angular-ui-router/release/angular-ui-router.js"></script>
+    <!-- <script src="./../node_modules/angular-ui-router/release/angular-ui-router.js"></script> -->
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/angular-ui-router/0.2.15/angular-ui-router.js"></script>
     <script src="app.js"></script>
-    <script src="postListController.js"></script>
+    <script src="./Controller/postListController.js"></script>
   </body>
 </html>
