--- postList.html
@@ -1,5 +1,0 @@
-<div ng-repeat='test in test'>
-  {{ test.username }}
-  {{ test.title }}
-  {{ test.message }}
-</div>
